+++ macros/src/lib.rs
@@ -400,6 +400,25 @@ use syn::{DeriveInput, Result};
 /// }
 /// # fn main() {}
 /// ```
+///
+/// # Overriding the path of the `aldrin` crate
+///
+/// The macro assumes per default that the `aldrin` crate is available as `::aldrin`. This can be
+/// overridden with the `krate` attribute. Note that [`generate!`] creates a new module and that
+/// path resolution starts inside that module.
+///
+/// ```
+/// # use aldrin_macros::generate;
+/// # fn main() {}
+/// mod my_reexports {
+///     pub use aldrin as my_aldrin;
+/// }
+///
+/// generate! {
+///     "schemas/example1.aldrin",
+///     crate = "super::my_reexports::my_aldrin",
+/// }
+/// ```
 #[manyhow::manyhow]
 #[proc_macro]
 pub fn generate(args: codegen::Args, emitter: &mut manyhow::Emitter) -> manyhow::Result {